Rename `storage` crate to `memory`
