+++ packages/engine/lib/memory/src/ffi/memory.rs
@@ -1,7 +1,7 @@
 #![allow(clippy::module_name_repetitions, clippy::missing_safety_doc)]
 use std::os::unix::io::RawFd;
 
-use crate::memory::Memory;
+use crate::shared_memory::Memory;
 
 #[repr(C)]
 pub struct CMemory {
