+++ packages/engine/lib/memory/src/meta/dynamic_meta.rs
