+++ packages/engine/lib/memory/src/meta/node.rs
