+++ packages/engine/lib/memory/src/meta/static_meta.rs
