+++ packages/engine/lib/memory/src/meta/util.rs
