+++ packages/engine/src/output/buffer/util.rs
@@ -1,6 +1,6 @@
 use std::path::PathBuf;
 
-use storage::memory::shmem_id_prefix;
+use storage::shared_memory::shmem_id_prefix;
 
 use crate::{
     output::{
