+++ packages/engine/src/worker/runner/python/fbs.rs
@@ -1,5 +1,5 @@
 use flatbuffers::{FlatBufferBuilder, WIPOffset};
-use storage::memory::{Metaversion, Segment};
+use storage::shared_memory::{Metaversion, Segment};
 
 use crate::{
     datastore::shared_store::SharedStore,
