+++ packages/engine/src/worker/runner/python/sender.rs
@@ -5,7 +5,7 @@ use arrow::{
 use flatbuffers::{FlatBufferBuilder, ForwardsUOffset, Vector, WIPOffset};
 use flatbuffers_gen::sync_state_interim_generated::StateInterimSyncArgs;
 use nng::{options::Options, Aio, Socket};
-use storage::memory::arrow_continuation;
+use storage::shared_memory::arrow_continuation;
 use tokio::sync::mpsc::{unbounded_channel, UnboundedReceiver};
 
 use crate::{
