+++ packages/engine/src/worker/runner/python/setup.py
@@ -30,7 +30,7 @@ def last_modified(path):
 
 
 library_dirs = sorted(
-    find_directory("libstorage.so", f"{script_path}/../../../../"),
+    find_directory("libmemory.so", f"{script_path}/../../../../"),
     key=last_modified,
     reverse=True,
 )
@@ -49,7 +49,7 @@ setup(
                 wrappers_import,
                 [script_path + "/wrappers.pyx"],
                 include_dirs=[numpy.get_include()],
-                libraries=["storage"],
+                libraries=["memory"],
                 library_dirs=library_dirs,
             )
         ]
